add utils and fix url-join
