+++ index.js
@@ -3,42 +3,7 @@ var fs = require('fs')
 var path = require('path')
 var Promise = require('bluebird')
 var OSS = require('ali-oss').Wrapper
-const unidecode = require('unidecode')
-function safeString (string, options) {
-  options = options || {};
-
-  if (string === null) {
-    string = '';
-  }
-
-  // Handle the £ symbol separately, since it needs to be removed before the unicode conversion.
-  string = string.replace(/£/g, '-');
-
-  // Remove non ascii characters
-  string = unidecode(string);
-
-  // Replace URL reserved chars: `@:/?#[]!$&()*+,;=` as well as `\%<>|^~£"{}` and \`
-  string = string.replace(/(\s|\.|@|:|\/|\?|#|\[|\]|!|\$|&|\(|\)|\*|\+|,|;|=|\\|%|<|>|\||\^|~|"|\{|\}|`|–|—)/g, '-')
-    // Remove apostrophes
-    .replace(/'/g, '')
-    // Make the whole thing lowercase
-    .toLowerCase();
-
-  // We do not need to make the following changes when importing data
-  if ((!options.hasOwnProperty('importing')) || !options.importing) {
-    // Convert 2 or more dashes into a single dash
-    string = string.replace(/-+/g, '-')
-      // Remove trailing dash
-      .replace(/-$/, '')
-      // Remove any dashes at the beginning
-      .replace(/^-/, '');
-  }
-
-  // Handle whitespace at the beginning or end.
-  string = string.trim();
-
-  return string;
-}
+var utils = require('./utils')
 
 var baseStore = require('ghost-storage-base')
 
@@ -62,7 +27,7 @@ class OssStore extends baseStore {
       .then(function (result) {
         // console.log(result)
         if(origin){
-          resolve(path.join(origin, result.name))
+          resolve(utils.joinUrl(origin, result.name))
         }else{
           resolve(result.url)
         }      
@@ -126,7 +91,7 @@ class OssStore extends baseStore {
       var name = path.basename(file.name, ext)
   
       if (keyOptions.safeString) {
-        name = safeString(name)
+        name = utils.safeString(name)
       }
   
       if (keyOptions.prefix) {
